move log into renderer
